typescript: root nodes page
